refactor: mail pakage 변경
